Solved AoC2020 Day 5 Part 1

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day05/Day05Test.kt
@@ -10,7 +10,7 @@ class Day05Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         BinaryBoarding(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
@@ -22,11 +22,11 @@ class Day05Test {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample1.txt", 820L),
-                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample2.txt", 357L),
-                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample3.txt", 567L),
-                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample4.txt", 119L),
-                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample5.txt", 820L)
+                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample1.txt", 820),
+                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample2.txt", 357),
+                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample3.txt", 567),
+                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample4.txt", 119),
+                Arguments.of("de/havox_design/aoc2020/day05/day05part1sample5.txt", 820)
             )
 
         @JvmStatic
